+++ homework/dhrod0325/03/src/components/Death/DeathTotalList.ts
@@ -2,8 +2,8 @@ import { DeathList } from './DeathList';
 import { DeathTotal } from './DeathTotal';
 import { Component } from '../../interfaces';
 import { Summary } from '../../types';
-import { useSpinner } from '../../lib/Spinner';
 import { api } from '../../lib/api';
+import { useSpinner } from '../Helper/Spinner';
 
 export class DeathTotalList implements Component {
   private readonly $total: DeathTotal;
